Refactor PlaysReplacements into a plugin form
