+++ osu!StreamCompanion.sln
@@ -37,6 +37,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FileMapDataSender", "plugin
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TcpSocketDataSender", "plugins\TcpSocketDataSender\TcpSocketDataSender.csproj", "{50E3D207-3812-433A-8F9A-757FF55C991A}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PlaysReplacements", "plugins\PlaysReplacements\PlaysReplacements.csproj", "{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -225,6 +227,18 @@ Global
 		{50E3D207-3812-433A-8F9A-757FF55C991A}.Remote debug|Any CPU.Build.0 = Release|Any CPU
 		{50E3D207-3812-433A-8F9A-757FF55C991A}.Remote debug|x86.ActiveCfg = Release|Any CPU
 		{50E3D207-3812-433A-8F9A-757FF55C991A}.Remote debug|x86.Build.0 = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Debug|x86.Build.0 = Debug|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Release|x86.ActiveCfg = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Release|x86.Build.0 = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Remote debug|Any CPU.ActiveCfg = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Remote debug|Any CPU.Build.0 = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Remote debug|x86.ActiveCfg = Release|Any CPU
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9}.Remote debug|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -241,6 +255,7 @@ Global
 		{0C50FF12-C298-424F-8B56-E65BCDDAB159} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
 		{32A61CD7-FA30-40CB-8A23-788BBD65D4A7} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
 		{50E3D207-3812-433A-8F9A-757FF55C991A} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
+		{3D20C5F7-1CFC-4313-8B21-D3FA298F82A9} = {7357CB17-E9D7-493F-BC1B-354F717D7AA3}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {90CD718D-2503-4BBC-867D-E8EF9C9DBC77}
